-   [BugFix]OnError Listener -   [BugFix] Bug is fixed when client want to set count for *GetContact* -   [Add] Async has *ASYNC_READY* state -   [Add] Add lastMessageId and firstMessageId to the *GetHistory* -   [Replace] Replace most of the *integer* params to *long*. -   [Removed] CHAT_READY state was removed from Async and added to Chat and state changes      to *CHAT_READY* when respond of the *getUserInfo* is received. -   [BugFix]  Add some field that added to the respond of the create thread -   [Add]  Now u can get instance unique id when u send text message.

+++ podchat/src/main/java/com/fanap/podchat/mainmodel/ChatMessageContent.java
@@ -13,7 +13,7 @@ public class ChatMessageContent {
     private boolean New;
     private int firstMessageId;
     private int lastMessageId;
-    private int offset;
+    private long offset;
     private String order;
     private List<Integer> threadIds;
 
@@ -33,11 +33,11 @@ public class ChatMessageContent {
         this.firstMessageId = firstMessageId;
     }
 
-    public int getOffset() {
+    public long getOffset() {
         return offset;
     }
 
-    public void setOffset(int offset) {
+    public void setOffset(long offset) {
         this.offset = offset;
     }
